Move ED64 X-series i2c into separate file; hook up to RTC subsystem
