+++ Makefile
@@ -44,7 +44,7 @@ LIBDRAGON_OBJS += \
 			 $(BUILD_DIR)/compress/shrinkler_dec_fast.o $(BUILD_DIR)/compress/shrinkler_dec.o \
 			 $(BUILD_DIR)/joybus.o $(BUILD_DIR)/joybus_accessory.o $(BUILD_DIR)/joybus_rtc.o $(BUILD_DIR)/pixelfx.o \
 			 $(BUILD_DIR)/joypad.o $(BUILD_DIR)/joypad_accessory.o \
-			 $(BUILD_DIR)/controller.o $(BUILD_DIR)/rtc.o $(BUILD_DIR)/rtc_internal.o \
+			 $(BUILD_DIR)/controller.o $(BUILD_DIR)/ed64x.o $(BUILD_DIR)/rtc.o $(BUILD_DIR)/rtc_internal.o \
 			 $(BUILD_DIR)/eeprom.o $(BUILD_DIR)/eepromfs.o $(BUILD_DIR)/mempak.o $(BUILD_DIR)/cpak.o \
 			 $(BUILD_DIR)/tpak.o $(BUILD_DIR)/graphics.o $(BUILD_DIR)/rdp.o \
 			 $(BUILD_DIR)/rsp.o $(BUILD_DIR)/rsp_crash.o \
